doce/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Semantic Kernel
semantic-kernel==1.28.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Google Vision
google-cloud-vision==3.4.5

# Azure OpenAI
openai
azure-identity==1.14.1

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
python-jose==3.3.0  # For JWT tokens
passlib==1.7.4      # For password hashing